Use class_create only with AJA_CREATE_DEVICE_NODES
Since class_create/destroy are GPL-only symbols, they are
not permitted to be used with a 5.9+ kernel when the
proprietary nvidia_p2p symbols are also used for RDMA
support. This change makes it so that class_create/destroy
are only called when AJA_CREATE_DEVICE_NODES is enabled,
allowing AJA_RDMA to be enabled for 5.9+ kernels. Note that
enabling both AJA_RDMA and AJA_CREATE_DEVICE_NODES on
5.9+ kernels is not yet supported.