Update nvidia-ko-to-module-symvers to support CRC table
CRC entries for symbols exported by kernel modules may be
provided either by absolute values in the symbol table, or
as offsets into the CRC table stored in a separate __kcrctab
data section. This change adds support for the latter type
of relative entry as it is needed to support RDMA with newer
kernels (>= 5.19).